requirements.txt
flask
Flask-SQLAlchemy
flask-login
flask-mail